dense_to_sparse: don't silently truncate
The width of a sparse matrix is implicitly defined by the largest column
value in it. It is assumed that there aren't any all-zero columns beyond
that width.

If the last column of a dense matrix is all zeroes, converting it to
sparse representation would silently truncate the width. One way around
this would be to store the width separately, but in the context of DLX
those matrices are not interesting so it would complicate the code for
no reason. Trying to convert such a matrix will just fail and return
null.